[SUS-32] Refactor
